+++ packages/client/src/pages/LeaderBoard/LeaderBoard.tsx
@@ -3,33 +3,7 @@ import { Box, Typography } from '@mui/material'
 import { leadersMock } from './mock'
 import back from '../../assets/chessboard-background.png'
 import './leaderboard.scss'
-
-interface IBlockLeader {
-  avatar: string
-  username: string
-  displayName: string
-  score: number | undefined
-  place: number
-}
-
-const LeaderBox: React.FC<IBlockLeader> = (props: IBlockLeader) => {
-  const { avatar, displayName, username, score } = props
-
-  return (
-    <Box className={'container-leaderbox'}>
-      <Box className={'container-leaderbox__left'}>
-        <Box className={'container-leaderbox__avatar'}>
-          <img src={avatar} alt="avatar" />
-        </Box>
-        <Box>
-          <Box>{displayName}</Box>
-          <small className="text-muted">{username}</small>
-        </Box>
-      </Box>
-      <Box className={'container-leaderbox__right'}>{score}</Box>
-    </Box>
-  )
-}
+import LeaderBox from './LeaderBox'
 
 const LeaderBoard: React.FC = () => (
   <Box
@@ -49,21 +23,16 @@ const LeaderBoard: React.FC = () => (
     </Typography>
 
     <Box className={'table-leaderboards'}>
-      {leadersMock.map((leader, index) => {
-        // Исходи из того что лидеры уже отсортированы на бэке по очкам
-        const currentPlace = index + 1
-
-        return (
-          <LeaderBox
-            key={index}
-            avatar={leader.avatar}
-            username={leader.username}
-            displayName={leader.displayName}
-            score={leader.score}
-            place={currentPlace}
-          />
-        )
-      })}
+      {leadersMock.map((leader, index) => (
+        <LeaderBox
+          key={`${leader.username}_${index}`}
+          avatar={leader.avatar}
+          username={leader.username}
+          displayName={leader.displayName}
+          score={leader.score}
+          place={index + 1}
+        />
+      ))}
     </Box>
   </Box>
 )
