+++ packages/client/src/pages/LeaderBoard/leaderboard.scss
@@ -14,7 +14,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  // background: `url(${back})`;
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
